requirements.txt
dnspython>=2.6.0
httpx>=0.27.0
jinja2>=3.1.4
rich>=13.7.0
python-whois>=0.9.4
